hipo4 compatibility updates to bank reading

+++ reconstruction/cnd/src/main/java/org/jlab/service/cnd/CNDCalibrationEngine.java
@@ -168,7 +168,7 @@ public class CNDCalibrationEngine extends ReconstructionEngine {
 	                        bank2.setFloat("y",i,   (float)(1.0* cndclusters.get(i).get_y()) );
 	                        bank2.setFloat("z",i,   (float)(1.0* cndclusters.get(i).get_z()) );
 	                        bank2.setFloat("time",i,   (float)(1.0*  cndclusters.get(i).get_time()) );
-				bank2.setInt("status",i,   cndclusters.get(i).get_status());
+				bank2.setShort("status",i, (short)  cndclusters.get(i).get_status());
 	                }
 	                event.appendBanks(bank2);
 	        }
